MAJOR CHANGES (see commit description!)
Major changes have been made to the core functionality of the game.   
- Character menu has been reworked; now includes a reset option.  

- Saving before quitting does not trigger suggestion to save; only if
the character did not save in the previous command or before the quit
statement in the current command will the save suggestion be triggered.  

- Draft for player death in the works; causes deletion of character save
to encourage players to avoid dying

- Taking has been redone; player may take directly from the environment
by simply including a direct object, or the player may take from a
container in the current environment, such as a locker, box, chest, etc.
This action makes use of the parser's prepositional phrase recognition
to determine the container of which to take from.  If a prepositional
phrase is not included, it is assumed that the player wants to take an
item from the environment

+++ game/reset.py
@@ -10,8 +10,17 @@ if os.name == 'posix':
     saves_dir = os.getcwd() + "/saves/"
 else:
     saves_dir = os.getcwd() + "\saves\\"
-    
-os.remove(saves_dir + 'Hunter' + '.dat')
+
+reset = False
+for file in os.listdir(saves_dir):
+    if file.endswith(".dat"):
+        if "hunter.dat" == file.lower():
+            reset = True
+            os.remove(saves_dir + 'Hunter' + '.dat')
+            
 game_defs.init_char('Hunter', saves_dir)
 
-print('Character reset.')
+if reset:
+    print('Hunter reset.'.format())
+else:
+    print('Hunter created.')
